requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
python-dotenv
wikipedia
numexpr